requirements.txt
PyYAML==3.11
amqp==1.4.5
anyjson==0.3.3
beautifulsoup4==4.3.2
billiard==3.3.0.17
celery==3.1.11
kombu==3.0.16
lxml==3.3.5
python-docx==0.5.1
pytz==2014.2
redis==2.9.1
